chatroom en cours

+++ app/controllers/api/v1/blocks_controller.rb
@@ -5,15 +5,17 @@ class Api::V1::BlocksController < ApplicationController
 
   def create
     blocked_user = User.find(params[:blocked_id])
-    current_user.blocks_given.create!(blocked: blocked_user)
+    block = current_user.blocks_given.create!(blocked: blocked_user)
     
+    authorize block
+
     # Supprimez également la discussion si nécessaire
     chat = Chatroom.find(params[:chatroom_id])
     chat.destroy
-        ChatroomChannel.broadcast_to(
-            chat,
-            { action: "chatroom_deleted" }
-        )
+      ChatroomChannel.broadcast_to(
+        chat,
+        { action: "chatroom_deleted" }
+      )
 
     # Supprimez du repertoire current et user
 
@@ -21,19 +23,19 @@ class Api::V1::BlocksController < ApplicationController
     repertoire_other = blocked_user.repertoire
     everyone_group_other = repertoire_other.contact_groups.find_by(name: 'Everyone')
     if everyone_group_other
-        association = UsersContactGroup.find_by(user: current_user, contact_group: everyone_group_other)
+        association = UserContactGroup.find_by(user: current_user, contact_group: everyone_group_other)
         association.destroy if association
     end
     # Repertoire du current_user
     repertoire = current_user.repertoire
     everyone_group = repertoire.contact_groups.find_by(name: 'Everyone')
     if everyone_group
-        association = UsersContactGroup.find_by(user: blocked_user, contact_group: everyone_group)
+        association = UserContactGroup.find_by(user: blocked_user, contact_group: everyone_group)
         association.destroy if association
     end
     
+    redirect_to api_v1_chatrooms_path(current_user), notice: "User has been blocked and the conversation was deleted."
 
-    redirect_to chatroom_path(current_user), notice: "User has been blocked and the conversation was deleted."
   end
 
   def destroy
